+++ src/app/components/connection-window/connection-window.component.ts
@@ -6,7 +6,7 @@ import { version } from '../../../../package.json';
 import { MysqlService } from '../../services/mysql.service';
 import { SubscriptionHandler } from '../../utils/subscription-handler/subscription-handler';
 import { Config } from '../../types/config.type';
-import { LocalStorageService } from '../../services/localstorage.service';
+import { LocalStorageService } from '../../services/local-storage.service';
 
 @Component({
   selector: 'app-connection-window',
